refactor: extract package for sub-domain
