+++ src/main/java/vendingmachine/model/coin/CoinsGenerator.java
@@ -1,4 +1,4 @@
-package vendingmachine.model;
+package vendingmachine.model.coin;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Arrays;
