+++ src/main/java/vendingmachine/reader/PurchaseItemNameReader.java
@@ -1,6 +1,6 @@
 package vendingmachine.reader;
 
-import vendingmachine.model.ItemRepository;
+import vendingmachine.model.item.ItemRepository;
 import vendingmachine.reader.validator.Validator;
 import vendingmachine.reader.validator.item.NotFoundItemValidator;
